+++ eslint.config.mjs
@@ -1,4 +1,17 @@
 // @ts-check
 import { defineConfig } from './dist/index.mjs';
+import globals from 'globals';
 
-export default defineConfig();
+export default defineConfig(
+  {
+    ignores: [
+      'examples/**',
+    ],
+  },
+  {
+    files: ['test/**/*.mjs'],
+    languageOptions: {
+      globals: globals.node,
+    },
+  },
+);