+++ examples/playground-nuxt/eslint.config.mjs
@@ -1,0 +1,5 @@
+// @ts-check
+import { forNuxt } from '@poupe/eslint-config/nuxt';
+import withNuxt from './.nuxt/eslint.config.mjs';
+
+export default withNuxt(...forNuxt());